updated create/delete projects flow
